Add fio to regression testing
fio - is a Flexible I/O tester [1]. Fio was originally written to test a
specific workload, either for performance reasons or to find/reproduce a
bug. It allows to test many possible workloads, see examples [2] and
recommended configurations to test latency and throughput, see [3] and
[4]. This patch adds wrapper to pytest around fio and a signle simple
configuration. Other configurations are coming.

1. https://fio.readthedocs.io/en/latest/fio_doc.html
2. https://github.com/axboe/fio/tree/master/examples
3. https://docs.oracle.com/en-us/iaas/Content/Block/References/samplefiocommandslinux.htm
4. https://arstechnica.com/gadgets/2020/02/how-fast-are-your-disks-find-out-the-open-source-way-with-fio/

Closes #51